requirements.txt
# Core dependencies
openai-agents>=0.1.0
pydantic>=2.0.0
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0
mypy>=1.5.0
coverage>=7.0.0

# Additional utilities
typing-extensions>=4.8.0
uuid>=1.30
